Refactored OpenBrowser out of Snapshot.
